feat: add key values and status bar render
